Finished constructors and added a few more bits where requried.
This whole project doesn't have a great structure due in main part
to not understanding the structure of an OSGi project. Everyone
in the team will need to go back and re-do the supplied lab
sheet which explains most of it in a semi-intelligible way.

Lots of work still to be done!